unified HmIPHeatingThermostat with HmIPWallMountedThermostat
