+++ src/HmIPPlatform.ts
@@ -1,21 +1,12 @@
-import {
-  API,
-  Characteristic,
-  DynamicPlatformPlugin,
-  Logger,
-  PlatformAccessory,
-  PlatformConfig,
-  Service,
-} from 'homebridge';
+import {API, Characteristic, DynamicPlatformPlugin, Logger, PlatformAccessory, PlatformConfig, Service} from 'homebridge';
 import {HmIPConnector} from './HmIPConnector';
 import {PLATFORM_NAME, PLUGIN_NAME, PLUGIN_VERSION} from './settings';
 import {HmIPDevice, HmIPGroup, HmIPHome, HmIPState, HmIPStateChange, Updateable} from './HmIPState';
 import {HmIPShutter} from './devices/HmIPShutter';
-import {HmIPWallMountedThermostat} from './devices/HmIPWallMountedThermostat';
+import {HmIPHeatingThermostat} from './devices/HmIPHeatingThermostat';
 import {HmIPContactSensor} from './devices/HmIPContactSensor';
 import {HmIPGenericDevice} from './devices/HmIPGenericDevice';
 import {HmIPAccessory} from './HmIPAccessory';
-import {HmIPHeatingThermostat} from './devices/HmIPHeatingThermostat';
 import * as os from 'os';
 import {HmIPSmokeDetector} from './devices/HmIPSmokeDetector';
 import {HmIPSwitch} from './devices/HmIPSwitch';
@@ -254,25 +245,23 @@ export class HmIPPlatform implements DynamicPlatformPlugin {
     const uuid = this.api.hap.uuid.generate(id);
     const hmIPAccessory = this.createAccessory(uuid, device.label, device);
     let homebridgeDevice: HmIPGenericDevice | null = null;
-    if (device.type === 'WALL_MOUNTED_THERMOSTAT_PRO'
-      || device.type === 'BRAND_WALL_MOUNTED_THERMOSTAT'
-      || device.type === 'ROOM_CONTROL_DEVICE'
-      || device.type === 'TEMPERATURE_HUMIDITY_SENSOR'
-      || device.type === 'TEMPERATURE_HUMIDITY_SENSOR_DISPLAY'
-      || device.type === 'WALL_MOUNTED_THERMOSTAT_BASIC_HUMIDITY') {
+    if (HmIPHeatingThermostat.isThermostat(device.type)) {
       const accessoryConfig = this.config['devices']?.[device.id];
       const asClimateSensor = accessoryConfig?.['asClimateSensor'] === true;
-      if (asClimateSensor) {
+      if (asClimateSensor && !HmIPHeatingThermostat.isHeatingThermostat(device.type)) {
         homebridgeDevice = new HmIPClimateSensor(this, hmIPAccessory.accessory);
       } else {
-        homebridgeDevice = new HmIPWallMountedThermostat(this, hmIPAccessory.accessory);
+        homebridgeDevice = new HmIPHeatingThermostat(this, hmIPAccessory.accessory);
       }
     } else if (device.type === 'TEMPERATURE_HUMIDITY_SENSOR_OUTDOOR') {
       homebridgeDevice = new HmIPClimateSensor(this, hmIPAccessory.accessory);
+      /*
     } else if (device.type === 'HEATING_THERMOSTAT'
       || device.type === 'HEATING_THERMOSTAT_COMPACT'
       || device.type === 'HEATING_THERMOSTAT_EVO') {
-      homebridgeDevice = new HmIPHeatingThermostat(this, hmIPAccessory.accessory);
+      homebridgeDevice = new HmIPLegacyHeatingThermostat(this, hmIPAccessory.accessory);
+
+       */
     } else if (device.type === 'FULL_FLUSH_SHUTTER'
       || device.type === 'BRAND_SHUTTER') {
       homebridgeDevice = new HmIPShutter(this, hmIPAccessory.accessory);
@@ -289,7 +278,7 @@ export class HmIPPlatform implements DynamicPlatformPlugin {
       homebridgeDevice = new HmIPRotaryHandleSensor(this, hmIPAccessory.accessory);
     } else if (device.type === 'SMOKE_DETECTOR') {
       homebridgeDevice = new HmIPSmokeDetector(this, hmIPAccessory.accessory);
-    } else if ( device.type === 'PLUGABLE_SWITCH'
+    } else if (device.type === 'PLUGABLE_SWITCH'
       || device.type === 'PRINTED_CIRCUIT_BOARD_SWITCH_BATTERY'
       || device.type === 'PRINTED_CIRCUIT_BOARD_SWITCH_2' // Only first channel
       || device.type === 'OPEN_COLLECTOR_8_MODULE' // Only first channel
@@ -297,7 +286,7 @@ export class HmIPPlatform implements DynamicPlatformPlugin {
       || device.type === 'WIRED_SWITCH_8' // Only first channel
       || device.type === 'DIN_RAIL_SWITCH_4') { // Only first channel
       homebridgeDevice = new HmIPSwitch(this, hmIPAccessory.accessory);
-    } else if ( device.type === 'PLUGABLE_SWITCH_MEASURING'
+    } else if (device.type === 'PLUGABLE_SWITCH_MEASURING'
       || device.type === 'BRAND_SWITCH_MEASURING'
       || device.type === 'FULL_FLUSH_SWITCH_MEASURING') {
       homebridgeDevice = new HmIPSwitchMeasuring(this, hmIPAccessory.accessory);
